refactor: PreviewGenerator 사용되는 패키지로 이동
